# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for pull requests managment in teams",
        "title": "PR Reviewer Assignment Service (Test Task, Fall 2025)",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/pullRequest/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PullRequests"
                ],
                "summary": "Создать PR и автоматически назначить до 2 ревьюверов из команды авторы",
                "parameters": [
                    {
                        "description": "Данные для создания",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/docs.CreatePRRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "PR создан",
                        "schema": {
                            "$ref": "#/definitions/docs.CreatePRResponse"
                        }
                    },
                    "401": {
                        "description": "Нет/неверный админский токен",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Автор/команда не найдены",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "PR уже существует",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pullRequest/merge": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PullRequests"
                ],
                "summary": "Пометить PR как MERGED (идемпотентная операция)",
                "parameters": [
                    {
                        "description": "Идентификатор PR",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/docs.MergePRRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PR в состоянии MERGED",
                        "schema": {
                            "$ref": "#/definitions/docs.MergePRResponse"
                        }
                    },
                    "401": {
                        "description": "Нет/неверный админский токен",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "PR не найден",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pullRequest/reassign": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PullRequests"
                ],
                "summary": "Переназначить конкретного ревьювера на другого из его команды",
                "parameters": [
                    {
                        "description": "Данные для переназначения",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/docs.ReassignRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Переназначение выполнено",
                        "schema": {
                            "$ref": "#/definitions/docs.ReassignResponse"
                        }
                    },
                    "401": {
                        "description": "Нет/неверный админский токен",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "PR или пользователь найден",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Нарушение доменных правил переназначения",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/users/getReview": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Получить PR'ы, где пользователь установлен ревьювером",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор пользователя",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список PR'ов пользователя",
                        "schema": {
                            "$ref": "#/definitions/docs.GetReviewResponse"
                        }
                    },
                    "401": {
                        "description": "Нет/неверный админский токен",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/setIsActive": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Установить флаг активности пользователя",
                "parameters": [
                    {
                        "description": "Данные для обновления",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/docs.SetIsActiveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленный пользователь",
                        "schema": {
                            "$ref": "#/definitions/docs.SetIsActiveResponse"
                        }
                    },
                    "401": {
                        "description": "Нет/неверный админский токен",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "docs.CreatePRRequest": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "pull_request_id": {
                    "type": "string"
                },
                "pull_request_name": {
                    "type": "string"
                }
            }
        },
        "docs.CreatePRResponse": {
            "type": "object",
            "properties": {
                "pr": {
                    "$ref": "#/definitions/docs.PRResponseObject"
                }
            }
        },
        "docs.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "docs.GetReviewPRResponse": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "pull_request_id": {
                    "type": "string"
                },
                "pull_request_name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "docs.GetReviewResponse": {
            "type": "object",
            "properties": {
                "pull_requests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/docs.GetReviewPRResponse"
                    }
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "docs.MergePRRequest": {
            "type": "object",
            "properties": {
                "pull_request_id": {
                    "type": "string"
                }
            }
        },
        "docs.MergePRResponse": {
            "type": "object",
            "properties": {
                "pr": {
                    "$ref": "#/definitions/docs.MergePRResponseObject"
                }
            }
        },
        "docs.MergePRResponseObject": {
            "type": "object",
            "properties": {
                "assigned_reviewers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "author_id": {
                    "type": "string"
                },
                "mergedAt": {
                    "type": "string"
                },
                "pull_request_id": {
                    "type": "string"
                },
                "pull_request_name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "docs.PRResponseObject": {
            "type": "object",
            "properties": {
                "assigned_reviewers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "author_id": {
                    "type": "string"
                },
                "pull_request_id": {
                    "type": "string"
                },
                "pull_request_name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "docs.ReassignRequest": {
            "type": "object",
            "properties": {
                "old_reviewer_id": {
                    "type": "string"
                },
                "pull_request_id": {
                    "type": "string"
                }
            }
        },
        "docs.ReassignResponse": {
            "type": "object",
            "properties": {
                "pr": {
                    "$ref": "#/definitions/docs.PRResponseObject"
                },
                "replaced_by": {
                    "type": "string"
                }
            }
        },
        "docs.SetIsActiveRequest": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "docs.SetIsActiveResponse": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean"
                },
                "team_name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}